Fix iOS 14 and Xcode 12 compilation issue

+++ CargoBay/CargoBay.m
@@ -284,8 +284,8 @@ BOOL CBValidateTransactionMatchesPurchaseInfo(SKPaymentTransaction *transaction,
     if (transactionQuantity != purchaseInfoDictionaryQuantity) {
         if (error) {
             NSDictionary *userInfo = [NSMutableDictionary dictionary];
-            [userInfo setValue:[NSString stringWithFormat:NSLocalizedStringFromTable(@"Transaction does not match purchase info because transaction's quantity (%d) does not match purchase info's quantity (%d).", @"CargoBay", nil), transactionQuantity, purchaseInfoDictionaryQuantity] forKey:NSLocalizedDescriptionKey];
-            [userInfo setValue:[NSString stringWithFormat:NSLocalizedStringFromTable(@"Transaction's quantity (%d) does not match purchase info's quantity (%d).", @"CargoBay", nil), transactionQuantity, purchaseInfoDictionaryQuantity] forKey:NSLocalizedFailureReasonErrorKey];
+            [userInfo setValue:[NSString stringWithFormat:NSLocalizedStringFromTable(@"Transaction does not match purchase info because transaction's quantity (%ld) does not match purchase info's quantity (%ld).", @"CargoBay", nil), (long)transactionQuantity, (long)purchaseInfoDictionaryQuantity] forKey:NSLocalizedDescriptionKey];
+            [userInfo setValue:[NSString stringWithFormat:NSLocalizedStringFromTable(@"Transaction's quantity (%ld) does not match purchase info's quantity (%ld).", @"CargoBay", nil), (long)transactionQuantity, (long)purchaseInfoDictionaryQuantity] forKey:NSLocalizedFailureReasonErrorKey];
             *error = [NSError errorWithDomain:CargoBayErrorDomain code:CargoBayErrorTransactionDoesNotMatchesPurchaseInfo userInfo:userInfo];
         }
